Modified SplashScreen
The SplashScreen was modified so that it depicts the logo and name of the application.
Images related to both the logo and name were added, and both subsequently transformed to "Sprite (2D and UI)" with "Single" as the Sprite mode.
The 2 images were merged in order to form a single element, which was also transformed like the previous ones. Finally, the latter was inserted as a SplashScreen image in combination with a background color.