Fixed SplashScreen
The previous SplashScreen had been implemented by Unity settings, those common to any project.
Now, however, those settings have been eliminated (making only the Unity logo appear) to make way for a new SplashScreen, implemented through a scene (the first one that will be started in the game.
The motivation for this lies in the fact that using the scene makes it possible to build the desired animations (as done here).